--- screens/AchievementsScreen.jsx
@@ -1,11 +1,0 @@
-import { View, Text } from 'react-native';
-
-const AchievementsScreen = () => {
-    return (
-        <View>
-            <Text>AchievementsScreen will go here</Text>
-        </View>
-      );
-}
- 
-export default AchievementsScreen;
